main/format_pack: Fix a wrong datatype in pack_ubyte_R8G8_UNORM
Before it was only storing one of the color components due to truncation.
With this patch it now properly stores all of them.

Cc: "10.2" <mesa-stable@lists.freedesktop.org>

+++ src/mesa/main/format_pack.c
@@ -888,7 +888,7 @@ pack_float_R_UNORM8(const GLfloat src[4], void *dst)
 static void
 pack_ubyte_R8G8_UNORM(const GLubyte src[4], void *dst)
 {
-   GLubyte *d = ((GLubyte *) dst);
+   GLushort *d = ((GLushort *) dst);
    *d = PACK_COLOR_88(src[GCOMP], src[RCOMP]);
 }
